fix: Riverpod 3.x and auto_route 9.x CI compilation errors
- Add StateProvider shim + valueOrNull extension to riverpod_compat
- Complete window_manager stub (TitleBarStyle, DragToMoveArea, setResizable, etc.)
- Add auto_route import to 17 music pages
- Fix initialRoutes removed in auto_route 9.x (use postFrameCallback)
- Fix const AutoRouter() not valid in auto_route 9.x
- Fix num->int cast in synced_lyrics.dart
- Fix FilePicker ambiguity (hide from shadcn_flutter)
- Fix items type cast in playlist_card.dart
- Add riverpod_compat import to metadata_plugin_provider
- Fix logger.dart base keyword

+++ lib/modules/music/pages/lyrics/synced_lyrics.dart
@@ -174,7 +174,7 @@ class SyncedLyrics extends HookConsumerWidget {
                                     onTap: () async {
                                       final time = Duration(
                                         seconds:
-                                            lyricSlice.time.inSeconds - delay,
+                                            (lyricSlice.time.inSeconds - delay).toInt(),
                                       );
                                       if (time > audioPlayer.duration ||
                                           time.isNegative) {
